refactor

+++ profile/Models/Crypto.cs
@@ -56,9 +56,9 @@ namespace profile.Models
         public int num_market_pairs { get; set; }
         public DateTime date_added { get; set; }
         //public List<object> tags { get; set; } list not supp, foreign key
-        public decimal? max_supply { get; set; }
-        public decimal? circulating_supply { get; set; }
-        public decimal? total_supply { get; set; }
+        public double? max_supply { get; set; }
+        public double? circulating_supply { get; set; }
+        public double? total_supply { get; set; }
         //public Platform platform { get; set; }
         public int cmc_rank { get; set; }
         public DateTime last_updated { get; set; }
@@ -72,12 +72,12 @@ namespace profile.Models
         [Required]
         [Key]
         public int usdid { get; set; }
-        public decimal? price { get; set; }
-        public decimal? volume_24h { get; set; }
-        public decimal? percent_change_1h { get; set; }
-        public decimal? percent_change_24h { get; set; }
-        public decimal? percent_change_7d { get; set; }
-        public decimal? market_cap { get; set; }
+        public Int64? price { get; set; }
+        public Int64? volume_24h { get; set; }
+        public Int64? percent_change_1h { get; set; }
+        public Int64? percent_change_24h { get; set; }
+        public Int64? percent_change_7d { get; set; }
+        public Int64? market_cap { get; set; }
         public DateTime last_updated { get; set; }
 
     }
